Modified the classes Utente and Liker to handle the many-to-many relation between the two classes.
Created the package compoundkeys that stores all the classes that represents the composed key
of all the many-to-many relations in the database, inside that there is the UserLikeComposedKey class
that represents the composed key of UserLiker the java class that is
the new entity created by the many-to-many relation between Utente and Liker.

Created the useless classes LikerController, LikeCrudRepository and LikerService
that have to be deleted in the next commit.

Changed the application.properties file because the new database is libra instead
of elis and because the ddl-auto now is setted on create-drop instead of update.